[INIT] Initialization of the JackOS version 2
I really enjoyed discovering OSDev with the first version of JackOS, and I hope you will enjoy this new version.
I will try to make clean code and clean commits, but I'm not a professional programmer, so I'm sorry if the code is not very clean.
The objective of this project is to learn how to make an OS, create the documentation
and try to transforms this project into a real OS (with a GUI, a shell, a file system, etc...).
I'm inspired by PeachOS, thanks to him for his tutorials.
(https://github.com/nibblebits/PeachOS)

Kernel (./src/kernel):
- Creation of the kernel.c that will handle kernel_main function that will be called by the bootloader
- Creation of the kernel.h that will handle the kernel_main function definition
and some handsome define likes VGA_WIDTH, VGA_HEIGHT and video_memory global variable
(it must be changed...).
- Creation of the entry_point.asm that will handle the entry point of the kernel.
I think I understand 80% of this file, I just stole it from PeachOS, I will
try to perfectly understand it later.

Boot (./src/boot):
- Creation of the boot.asm that will handle the bootloader. I don't understand
a percent of this file, it's likes magic for me. I just stole it from PeachOS,
trying to understand it later.

Other:
- Creation of a gitignore that ignores the build files
- Creation of a Makefile that compiles the kernel and the bootloader with
many features
- Creation of the bin folder that will handle all binaries files, it contains
.gitkeep file to keep the folder in the repository
- Creation of the build folder that will handle all build files (.o), it contains
.gitkeep file to keep the folder in the repository
- Creation of the ./src/linker.ld that will handle the linker script