[ADD] Creating of first heap system
The heap system is necessary when you want to have dynamic memory.
I created memory heap management to open possibilities of dynamic memory
allocation. The heap system is based on the first-fit algorithm.

Other:
- Adding of file in Makefile to compile:
  src/heap/heap.c
- Creating of src/heap folder that will contains all heap files and functions
- Adding of heap initialization in kernel_main function (src/kernel.c)

Heap (src/heap):
- Creating of heap.c file that contains all functions to manage the heap
It contains for now only the heap_init function that initialize the heap
and its data table and check the aligment, it's raise panic if the aligment
is not correct (HEAP_BLOCK_SIZE).
Adding of functions:
- void heap_init();